Add data controller

+++ E-Centrar-API/ServiceLayers/DTOs/CustomerDTO.cs
@@ -23,8 +23,8 @@ namespace ServiceLayers.DTOs
         public string UpdatedBy { get; set; }
         public string PaymentMethod { get; set; }
         public int FK_RouteId { get; set; }
-        public float Latitude { get; set; }
-        public float Longitude { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
 
 
     }
